Refactor graphql_helpers tests into the parsers tests, add tests for all the parser classes
